% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_detect_dummy_cols.R
\name{clean_detect_dummy_cols}
\alias{clean_detect_dummy_cols}
\title{Detect if a dataframe has dummy columns}
\usage{
clean_detect_dummy_cols(.data)
}
\arguments{
\item{.data}{a dataframe}
}
\value{

}
\description{
Detect if a dataframe has dummy columns
}
\examples{
.data <- tibble(test = 1:5,
                to_dummy = c('level1', 'level1', 'level2', 'level2', 'level3'),
                to_dummy2 = c('char1', 'char3', 'char3', 'char2', 'char1'))
.data <- fastDummies::dummy_cols(.data)
clean_detect_dummy_cols(.data)
}
\author{
Joe Marlo
}
