% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_help.R
\name{create_link_to_help}
\alias{create_link_to_help}
\title{Create a link to the help section}
\usage{
create_link_to_help(
  header,
  button_label = "Help me",
  type = c("button", "run_js", "js_code")
)
}
\arguments{
\item{header}{The h3 header text within help.md to jump to}

\item{button_label}{Label text to included on the button}

\item{type}{one of c('button', 'run_js', 'js_code'). 'button' returns the html for a button. 'js_code' returns just the string of JavaScript code. 'run_js' runs the JavaScript.}
}
\value{

}
\description{
Creates a link to open the help slide-over to the header provided. Defaults to returning a button to be pressed.
}
\examples{
# within shiny UI
create_link_to_help('Data')

# within shiny server
create_link_to_help('Data', type = 'run_js')
}
