% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_interactive_table.R
\name{get_table_values}
\alias{get_table_values}
\title{Retrieve user inputted values from a table}
\usage{
get_table_values(input, table_id, ns = NULL, convert_to_numeric = TRUE)
}
\arguments{
\item{input}{the global `input` object within Shiny server}

\item{table_id}{string. The id assigned to the table when creating the table}

\item{ns}{optional. A function defining the name space. To be used with shiny::NS}

\item{convert_to_numeric}{defaults to TRUE. Automatically coerces values to numeric}
}
\value{

}
\description{
Function to use on the server side to retreive the user input from an interactive table.
}
\examples{
.data <- tibble(
  Student = c('Blake', 'Kennedy', 'Jordan', 'Jordan', 'John'), 
  Z = c(1, 1, 0, 0, 0),
  Y1 = c(115, 130, '?', '?', 120),
  Y0 = c('?', '?', 130, 128, 130),
  Y = c(115, 130, 130, 128, 130),
  ITE = c(-6, -5, -2, -8, "?")
 )
correct_answers <- c("121", "135", "127", "120", "5")
html_string <- create_interactive_table(.data, correct_answers, table_id = 'mytable')
# HTML(html_string) within Shiny UI 
# get_table_values(input, 'mytable') within Shiny server
}
\seealso{
\code{\link{create_interactive_table}}
}
\author{
Joe Marlo
}
