% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ns_functions.R
\name{make_namespace_function}
\alias{make_namespace_function}
\title{Closure to create namespace functions}
\usage{
make_namespace_function(module_id)
}
\arguments{
\item{module_id}{the id of the namespace}
}
\value{
a function similar to shiny::ns
}
\description{
This function creates other functions which help with namespace ids. This is similar to shiny::NS but it can be applied programmatically to create many functions at once.
}
\examples{
make_namespace_function("my_module_id")
ns_my_module_id("input_id")
}
